Refactor bot launchers classes
