+++ src/main/java/com/botica/launchers/TestCaseExecutorLauncher.java
@@ -5,7 +5,6 @@ import static es.us.isa.restest.util.FileManager.deleteDir;
 
 import org.json.JSONObject;
 
-import com.botica.bots.TestCaseExecutor;
 import com.botica.utils.BotConfig;
 import com.botica.utils.RESTestUtil;
 
@@ -14,10 +13,23 @@ import es.us.isa.restest.runners.RESTestExecutor;
 /**
  * This class is a launcher for executing test cases generated by generators.
  */
-public class TestCaseExecutorLauncher extends BaseLauncher{
+public class TestCaseExecutorLauncher extends AbstractLauncher{
 
     private static final String BOT_TYPE = "testCaseExecutor";
     private static final String BINDING_KEY = "testCasesGenerated";
+
+    private String propertyFilePath;
+    private String testCasesPath;
+
+    public TestCaseExecutorLauncher(String keyToPublish, String orderToPublish) {
+        super(keyToPublish, orderToPublish);
+    }
+
+    public TestCaseExecutorLauncher(String propertyFilePath, String testCasesPath, String keyToPublish, String orderToPublish) {
+        super(keyToPublish, orderToPublish);
+        this.propertyFilePath = propertyFilePath;
+        this.testCasesPath = testCasesPath;
+    }
 
     /**
      * Launches test case executor based on bot data provided, and sends and 
@@ -28,26 +40,19 @@ public class TestCaseExecutorLauncher extends BaseLauncher{
      * @param keyToPublish      The binding key for publishing messages to RabbitMQ.
      * @param orderToPublish    The order to send in the message.
      */
-    public void launchTestExecutor(JSONObject botData, String order, String keyToPublish, String orderToPublish) {
+    public void launchTestExecutor(JSONObject botData, String order) {
         
-        BotConfig botConfig = new BotConfig(null, order, keyToPublish, orderToPublish, BOT_TYPE);
+        BotConfig botConfig = new BotConfig(null, order, this.keyToPublish, this.orderToPublish, BOT_TYPE);
         String queueName = BOT_TYPE;
         launchBot(botData, botConfig, queueName, BINDING_KEY, false);
     }
 
     /**
      * Executes test cases generated.
-     *
-     * @param propertyFilePath The path to the property file.
-     * @param testCasesPath    The path to the test cases generated.
-     * @param keyToPublish     The binding key for publishing messages to RabbitMQ.
-     * @param orderToPublish   The order to send in the message.
      */
-    public static void executeTestCases(String propertyFilePath, String testCasesPath, String keyToPublish, String orderToPublish) {
-
+    @Override
+    protected void botAction() {
         RESTestExecutor executor = new RESTestExecutor(propertyFilePath);
-
-        TestCaseExecutor testExecutor = new TestCaseExecutor(executor, propertyFilePath, testCasesPath, keyToPublish, orderToPublish);
 
         // TODO: Check if is correct
         // Create directories to store test data extracted from the execution
@@ -70,6 +75,23 @@ public class TestCaseExecutorLauncher extends BaseLauncher{
         createDir(coverageDataDir);
         //
 
-        testExecutor.executeBotActionAndSendMessage();
+        auxBotAction(executor);
+    }
+
+    private void auxBotAction(RESTestExecutor executor) {
+        String allureResultsDirPath = RESTestUtil.readProperty(propertyFilePath, "allure.results.dir");
+        String experimentName = RESTestUtil.readProperty(propertyFilePath, "experiment.name");
+        System.setProperty("allure.results.directory", allureResultsDirPath + "/" + experimentName);
+        executor.execute();
+    }
+
+    @Override
+    protected JSONObject createMessage() {
+        JSONObject message = new JSONObject();
+        message.put("order", orderToPublish);
+        message.put("propertyFilePath", propertyFilePath);
+        message.put("testCasesPath", testCasesPath);
+
+        return message;
     }
 }
