+++ src/main/java/com/botica/runners/TestCaseExecution.java
@@ -14,8 +14,8 @@ public class TestCaseExecution extends RunnerBase{
         try {
             JSONObject testCaseExecutors = loadBotsDefinition(JSON_OBJECT);
             launchBots(testCaseExecutors, (botData, o, k, op) -> {
-                TestCaseExecutorLauncher launcher = new TestCaseExecutorLauncher();
-                launcher.launchTestExecutor(botData, o, k, op);
+                TestCaseExecutorLauncher launcher = new TestCaseExecutorLauncher(k, op);
+                launcher.launchTestExecutor(botData, o);
             });
         } catch (JSONException e) {
             throw new JSONException("Error reading file: " + DEFAULT_BOTS_DEFINITION_PATH);
