+++ src/main/java/com/botica/runners/TestReportGeneration.java
@@ -15,8 +15,8 @@ public class TestReportGeneration extends RunnerBase {
             JSONObject testReporters = loadBotsDefinition(JSON_OBJECT);
 
             launchBots(testReporters, (botData, o, k, op) -> {
-                TestReportGeneratorLauncher launcher = new TestReportGeneratorLauncher();
-                launcher.launchTestReportGenerator(botData, o, k, op);
+                TestReportGeneratorLauncher launcher = new TestReportGeneratorLauncher(k, op);
+                launcher.launchTestReportGenerator(botData, o);
             });
         } catch (JSONException e) {
             throw new JSONException("Error reading file: " + DEFAULT_BOTS_DEFINITION_PATH);
